+++ ui/__init__.py
@@ -1,7 +1,7 @@
 ui_modules = [
     "color_def", "sv_IO_panel", "sv_examples_menu", 'sv_3d_panel', 'nodeview_operators',
     "sv_panels", "nodeview_rclick_menu", "nodeview_space_menu","nodeview_solids_menu",
-    "sv_icons", "presets", "nodes_replacement", "sv_panel_display_nodes",
+    "sv_icons", "presets", "presets_operators", "nodes_replacement", "sv_panel_display_nodes",
     "sv_temporal_viewers", "sv_vep_connector", "zoom_to_node",
     # bgl modules
     "bgl_callback_3dview", "bgl_callback_nodeview",
